OTAT-552: Add view list of space assessments service with tests

+++ src/main/java/org/flickit/flickitassessmentcore/domain/Assessment.java
@@ -20,7 +20,7 @@ public class Assessment {
     private LocalDateTime creationTime;
     private LocalDateTime lastModificationDate;
     private Long assessmentKitId;
-    private Long colorId;
+    private Integer colorId;
     private Long spaceId;
     private List<AssessmentResult> assessmentResults;
     private List<Evidence> evidences;
